+++ NEWS.md
@@ -1,6 +1,7 @@
 # resourcecode (development version)
 
-- Custom labels are now handeled properly in `cut_seasons()` 
+- Custom labels are now handled properly in `cut_seasons()` ;
+- Switch to {httr2} to download files and API request.
 
 # resourcecode 0.5.2